// src/app/entities/permiso.enum.ts
export enum Permiso {
  CONSULTA_CLIENTE = 1,
  ACTUALIZAR_CLIENTES = 2,
  AGREGAR_CLIENTE = 3,
  ELIMINAR_CLIENTE = 4,

  CONSULTAR_USUARIO = 5,
  AGREGAR_USUARIO = 6,
  ACTUALIZAR_USUARIO = 7,
  ELIMINAR_USUARIO = 8,

  AGREGAR_MODULO = 9,
  ACTUALIZAR_MODULO = 10,
  CONSULTAR_MODULO = 11,
  ELIMINAR_MODULO = 12,

  AGREGAR_PERMISO = 13,
  ACTUALIZAR_PERMISO = 14,
  CONSULTAR_PERMISOS = 15,
  ELIMINAR_PERMISO = 16,

  AGREGAR_MARCA = 17,
  ACTUALIZAR_MARCAS = 18,
  CONSULTAR_MARCAS = 19,
  ELIMINAR_MARCA = 20,

  AGREGAR_MODELO = 21,
  ACTUALIZAR_MODELO = 22,
  CONSULTAR_MODELOS = 23,
  ELIMINAR_MODELO = 24,

  AGREGAR_EQUIPO = 25,
  ACTUALIZAR_EQUIPO = 26,
  CONSULTAR_EQUIPOS = 27,
  ELIMINAR_EQUIPO = 28,

  CONSULTAR_BITACORA = 29,

  AGREGAR_ROL = 30,
  ACTUALIZAR_ROL = 31,
  CONSULTAR_ROLES = 32,
  ELIMINAR_ROL = 33,

  CONSULTAR_MONITOREO = 34,

  AGREGAR_OFICINA_CENTRAL = 35,
  ACTUALIZAR_OFICINA_CENTRAL = 36,
  CONSULTAR_OFICINA_CENTRAL = 37,
  ELIMINAR_OFICINA_CENTRAL = 38,

  CONSULTAR_INCIDENCIAS = 39,

  CONSULTAR_TABLERO = 40,

  CONSULTAR_PRODUCTOS = 41,
  ACTUALIZAR_PRODUCTOS = 42,
  AGREGAR_PRODUCTO = 43,
  ELIMINAR_PRODUCTO = 44,

  CONSULTAR_DEPARTAMENTOS = 45,
  ACTUALIZAR_DEPARTAMENTOS = 46,
  AGREGAR_DEPARTAMENTOS = 47,
  ELIMINAR_DEPARTAMENTOS = 48,

  CONSULTAR_INSTALACIONES_GRID = 49,
  ACTUALIZAR_INSTALACIONES_GRID = 50,
  AGREGAR_INSTALACION_GRID = 51,
  ELIMINAR_INSTALACION_GRID = 52,

  CONSULTAR_INCREMENTOS = 53,
  ACTUALIZAR_INCREMENTOS = 54,
  AGREGAR_INCREMENTOS = 55,
  ELIMINAR_INCREMENTOS = 56,

  CONSULTAR_FACTORES = 57,
  ACTUALIZAR_FACTORES = 58,
  AGREGAR_FACTORES = 59,
  ELIMINAR_FACTORES = 60,
}